+++ ydb/_topic_writer/topic_writer.py
@@ -303,7 +303,7 @@ def messages_to_proto_requests(
         req = StreamWriteMessage.FromClient(
             StreamWriteMessage.WriteRequest(
                 messages=[msg.to_message_data()],
-                codec=Codec.CODEC_RAW.value,
+                codec=Codec.CODEC_RAW,
             )
         )
         res.append(req)