+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add control plane operations for topic api: create, drop
+
 ## 3.0.1b4 ##
 * Initial implementation of topic reader